+++ backend/bang-ggood/src/main/java/com/bang_ggood/auth/service/AuthService.java
@@ -1,15 +1,10 @@
 package com.bang_ggood.auth.service;
 
-import com.bang_ggood.auth.domain.PasswordResetCode;
-import com.bang_ggood.auth.dto.request.ConfirmPasswordResetCodeRequest;
-import com.bang_ggood.auth.dto.request.ForgotPasswordRequest;
 import com.bang_ggood.auth.dto.request.LocalLoginRequestV1;
 import com.bang_ggood.auth.dto.request.OauthLoginRequest;
 import com.bang_ggood.auth.dto.request.RegisterRequestV1;
-import com.bang_ggood.auth.dto.request.ResetPasswordRequest;
 import com.bang_ggood.auth.dto.response.AuthTokenResponse;
 import com.bang_ggood.auth.dto.response.OauthInfoApiResponse;
-import com.bang_ggood.auth.repository.PasswordResetCodeRepository;
 import com.bang_ggood.auth.service.jwt.JwtTokenProvider;
 import com.bang_ggood.auth.service.jwt.JwtTokenResolver;
 import com.bang_ggood.auth.service.oauth.OauthClient;
@@ -18,35 +13,25 @@ import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
 import com.bang_ggood.user.domain.Email;
 import com.bang_ggood.user.domain.LoginType;
-import com.bang_ggood.user.domain.Password;
 import com.bang_ggood.user.domain.User;
 import com.bang_ggood.user.domain.UserType;
 import com.bang_ggood.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.time.Clock;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @RequiredArgsConstructor
 @Service
 public class AuthService {
 
-    private static final Logger log = LoggerFactory.getLogger(AuthService.class);
     private static final int GUEST_USER_LIMIT = 1;
-    private static final int PASSWORD_RESET_CODE_EXPIRED_MINUTES = 5;
 
     private final OauthClient oauthClient;
     private final JwtTokenProvider jwtTokenProvider;
     private final JwtTokenResolver jwtTokenResolver;
     private final DefaultChecklistService defaultChecklistService;
     private final UserRepository userRepository;
-    private final MailSender mailSender;
-    private final PasswordResetCodeRepository passwordResetCodeRepository;
-    private final Clock clock;
 
     @Transactional
     public Long register(RegisterRequestV1 request) {
@@ -138,34 +123,6 @@ public class AuthService {
         if (!accessAuthUser.id().equals(refreshAuthUser.id())) {
             throw new BangggoodException(ExceptionCode.AUTHENTICATION_TOKEN_USER_MISMATCH);
         }
-    }
-
-    public void sendPasswordResetEmail(ForgotPasswordRequest request) {
-        String code = mailSender.sendPasswordResetEmail(request.email());
-        passwordResetCodeRepository.save(new PasswordResetCode(request.email(), code));
-    }
-
-    @Transactional(readOnly = true)
-    public void confirmPasswordResetCode(ConfirmPasswordResetCodeRequest request) {
-        LocalDateTime timeLimit = LocalDateTime.now(clock).minusMinutes(PASSWORD_RESET_CODE_EXPIRED_MINUTES);
-        boolean isValid = passwordResetCodeRepository.existsByEmailAndCodeAndCreatedAtAfter(
-                new Email(request.email()), request.code(), timeLimit);
-        if (!isValid) {
-            throw new BangggoodException(ExceptionCode.AUTHENTICATION_PASSWORD_CODE_NOT_FOUND);
-        }
-    }
-
-    @Transactional
-    public void resetPassword(ResetPasswordRequest request) {
-        Email email = new Email(request.email());
-        String code = request.code();
-        if (!passwordResetCodeRepository.existsByEmailAndCode(email, code)) {
-            throw new BangggoodException(ExceptionCode.AUTHENTICATION_PASSWORD_CODE_NOT_FOUND);
-        }
-
-        userRepository.updatePasswordByEmail(
-                new Email(request.email()), new Password(request.newPassword()), LoginType.LOCAL);
-        passwordResetCodeRepository.deleteByEmailAndCode(email, code);
     }
 
     @Transactional(readOnly = true)
